+++ app/components/views/GetStartedPage/Settings/Form.js
@@ -1,7 +1,9 @@
-import GeneralSettings from "views/SettingsPage/GeneralSettings";
-import PrivacySettings from "views/SettingsPage/PrivacySettings";
+import NetworkSettings from "views/SettingsPage/NetworkSettings";
 import ProxySettings from "views/SettingsPage/ProxySettings";
-import UIThemeSettings from "views/SettingsPage/UIThemeSettings";
+import PrivacySettings from "views/SettingsPage/PrivacySettings";
+import UISettings from "views/SettingsPage/UISettings";
+import MiscSettings from "views/SettingsPage/MiscSettings";
+import TimezoneSettings from "views/SettingsPage/TimezoneSettings";
 import { Tooltip, Subtitle } from "shared";
 import { FormattedMessage as T } from "react-intl";
 import { LoaderBarBottom } from "indicators";
@@ -12,7 +14,6 @@ import cx from "classnames";
 export default ({
   areSettingsDirty,
   tempSettings,
-  networks,
   currencies,
   locales,
   onChangeTempSettings,
@@ -24,8 +25,11 @@ export default ({
   getEstimatedTimeLeft,
   appVersion,
   updateAvailable,
-  getWalletReady,
-  isTestNet
+  isTestNet,
+  walletReady,
+  onAttemptChangePassphrase,
+  isChangePassPhraseDisabled,
+  changePassphraseRequestAttempt
 }) => (
   <div className={cx("page-body getstarted", isTestNet && "testnet-body")}>
     <div className="getstarted loader logs settings">
@@ -43,13 +47,48 @@ export default ({
       </div>
       <Subtitle title={<T id="settings.subtitle" m="Settings"/>} />
       <div className="settings-wrapper">
-        <div className="settings-columns">
-          <GeneralSettings {...{ tempSettings, networks, currencies, locales, onChangeTempSettings }} walletReady={getWalletReady}/>
-          <ProxySettings {...{ tempSettings, onChangeTempSettings }} />
-          <UIThemeSettings {...{ tempSettings, onChangeTempSettings }} />
+        <div className="settings-group">
+          <div className="settings-group-title"><T id="settings.group-title.connectivity" m="Connectivity" /></div>
+          <div className="settings-column-wrapper">
+            <div className="settings-column">
+              <NetworkSettings {...{
+                tempSettings,
+                onChangeTempSettings
+              }} />
+            </div>
+            <div className="settings-column">
+              <ProxySettings {...{ tempSettings, onChangeTempSettings }} />
+            </div>
+          </div>
         </div>
-        <div className="settings-columns">
-          <PrivacySettings {...{ tempSettings, onChangeTempSettings }} />
+
+        <div className="settings-group general">
+          <div className="settings-group-title"><T id="settings.group-title.general" m="General" /></div>
+          <div className="settings-column-wrapper">
+            <div className="settings-column">
+              <UISettings {...{ tempSettings, locales, onChangeTempSettings }} />
+            </div>
+            <div className="settings-column timezone">
+              <TimezoneSettings {...{ tempSettings, onChangeTempSettings }} />
+            </div>
+            {walletReady &&
+              <div className="settings-column">
+                <MiscSettings {...{ tempSettings, currencies, walletReady, onChangeTempSettings }} />
+              </div>
+            }
+          </div>
+        </div>
+
+        <div className="settings-group privacy">
+          <div className="settings-group-title"><T id="settings.group-title.privacy-and-security" m="Privacy and Security" /></div>
+          <div className="settings-column-wrapper">
+            <div className="settings-column">
+              <PrivacySettings {...{
+                tempSettings, onAttemptChangePassphrase,
+                isChangePassPhraseDisabled, onChangeTempSettings, changePassphraseRequestAttempt
+              }} />
+            </div>
+          </div>
         </div>
       </div>
       <div className="save-button-wrapper">
