+++ app/components/views/SettingsPage/PrivacySettings.js
@@ -6,6 +6,18 @@ import {
   EXTERNALREQUEST_DCRDATA
 } from "main_dev/externalRequests";
 
+import { ChangePassphraseButton } from "buttons";
+import { WatchOnlyWarnNotification } from "shared";
+
+
+const propTypes = {
+  tempSettings: PropTypes.object.isRequired,
+  onAttemptChangePassphrase: PropTypes.func,
+  isChangePassPhraseDisabled: PropTypes.bool.isRequired,
+  changePassphraseRequestAttempt: PropTypes.bool.isRequired,
+  onChangeTempSettings: PropTypes.func.isRequired
+};
+
 const AllowableRequestType = ({ id, label, description, checked, onChange }) => (
   <div className="settings-row settings-row-checklist">
     <div className="settings-label">
@@ -23,6 +35,9 @@ const AllowableRequestType = ({ id, label, description, checked, onChange }) => 
 
 const PrivacySettings = ({
   tempSettings,
+  isChangePassPhraseDisabled,
+  changePassphraseRequestAttempt,
+  onAttemptChangePassphrase,
   onChangeTempSettings
 }) => {
   const toggle = (value) => () => {
@@ -38,22 +53,22 @@ const PrivacySettings = ({
 
   return (
     <div className="settings-privacy">
-      <div className="settings-column-title"><T id="settings.privacy.title" m="Privacy" /></div>
       <div className="settings-column-content">
-        <AllowableRequestType
-          label={<T id="settings.privacy.networkStatus.label" m="Network Information" />}
-          id="networking"
-          description={<T id="settings.privacy.networkStatus.description" m="General network information (block height, etc) from decred.org" />}
-          checked={tempSettings.allowedExternalRequests.indexOf(EXTERNALREQUEST_NETWORK_STATUS) > -1}
-          onChange={toggle(EXTERNALREQUEST_NETWORK_STATUS)}
-        />
-        <AllowableRequestType
-          label={<T id="settings.privacy.stakepoolListing.label" m="VSP Listing" />}
-          id="stakepool"
-          description={<T id="settings.privacy.stakepoolListing.description" m="List of currently available VSPs from decred.org" />}
-          checked={tempSettings.allowedExternalRequests.indexOf(EXTERNALREQUEST_STAKEPOOL_LISTING) > -1}
-          onChange={toggle(EXTERNALREQUEST_STAKEPOOL_LISTING)}
-        />
+        <div className="settings-row settings-row-checklist">
+          <div disabled={isChangePassPhraseDisabled} className="settings-update-passphrase-button">
+            <T id="settings.updatePrivatePassphrase" m="Update Private Passphrase" />
+            <WatchOnlyWarnNotification isActive={isChangePassPhraseDisabled}>
+              <ChangePassphraseButton
+                className={[
+                  isChangePassPhraseDisabled ? "change-password-disabled-icon" : "",
+                  changePassphraseRequestAttempt ? "change-password-loading" : "" ].join(" ")}
+                isDisabled={isChangePassPhraseDisabled}
+                modalTitle={<T id="settings.changeConfirmation" m="Change your passphrase" />}
+                onSubmit={onAttemptChangePassphrase} />
+            </WatchOnlyWarnNotification>
+          </div>
+        </div>
+
         <AllowableRequestType
           label={<T id="settings.privacy.updateCheck.label" m="Update Check" />}
           id="update"
@@ -62,11 +77,25 @@ const PrivacySettings = ({
           onChange={toggle(EXTERNALREQUEST_UPDATE_CHECK)}
         />
         <AllowableRequestType
+          label={<T id="settings.privacy.networkStatus.label" m="Network Information" />}
+          id="networking"
+          description={<T id="settings.privacy.networkStatus.description" m="General network information (block height, etc) from decred.org" />}
+          checked={tempSettings.allowedExternalRequests.indexOf(EXTERNALREQUEST_NETWORK_STATUS) > -1}
+          onChange={toggle(EXTERNALREQUEST_NETWORK_STATUS)}
+        />
+        <AllowableRequestType
           label={<T id="settings.privacy.politeia.label" m="Politeia" />}
           id="politeia"
           description={<T id="settings.privacy.politeia.description" m="List and vote on proposals on proposals.decred.org" />}
           checked={tempSettings.allowedExternalRequests.indexOf(EXTERNALREQUEST_POLITEIA) > -1}
           onChange={toggle(EXTERNALREQUEST_POLITEIA)}
+        />
+        <AllowableRequestType
+          label={<T id="settings.privacy.stakepoolListing.label" m="VSP Listing" />}
+          id="stakepool"
+          description={<T id="settings.privacy.stakepoolListing.description" m="List of currently available VSPs from decred.org" />}
+          checked={tempSettings.allowedExternalRequests.indexOf(EXTERNALREQUEST_STAKEPOOL_LISTING) > -1}
+          onChange={toggle(EXTERNALREQUEST_STAKEPOOL_LISTING)}
         />
         <AllowableRequestType
           label={<T id="settings.privacy.dcrdata.label" m="Decred Block Explorer" />}
@@ -79,5 +108,7 @@ const PrivacySettings = ({
     </div>
   );
 };
+
+PrivacySettings.propTypes = propTypes;
 
 export default PrivacySettings;
