+++ app/style/GetStarted.less
@@ -99,6 +99,9 @@
     font-size: 27px;
     margin-left: 10px;
   }
+  .getstarted.settings .tabbed-page-subtitle {
+    margin-left: 0;
+  }
 }
 
 .getstarted {
@@ -244,12 +247,12 @@
   &.settings {
     .go-back-screen-button-area {
       position: absolute;
-      left: 710px;
+      left: 1467px;
       top: 112px;
       right: initial;
     }
     .save-button-wrapper {
-        width: 740px;
+        width: 1484px;
         text-align: right;
         .key-blue-button {
           margin-top: 0;
@@ -937,10 +940,43 @@
   margin-top: 14em;
 }
 
-@media screen and (max-width: 1106px) {
+@media screen and (max-width: 1919px) {
   .getstarted.loader.settings {
     .go-back-screen-button-area {
-      left: 620px;
+      left: 1230px;
+    }
+    .save-button-wrapper {
+      width: 1247px;
+    } 
+  }            
+}
+
+@media screen and (max-width: 1679px) {
+  .getstarted.loader.settings {
+    .go-back-screen-button-area {
+      left: 988px;
+    }
+    .save-button-wrapper {
+      width: 1002px;
+    } 
+  }            
+}
+
+@media screen and (max-width: 1439px) {
+  .getstarted.loader.settings {
+    .go-back-screen-button-area {
+      left: 726px;
+    }
+    .save-button-wrapper {
+      width: 740px;
+    } 
+  }            
+}
+
+@media screen and (max-width: 1180px) {
+  .getstarted.loader.settings {
+    .go-back-screen-button-area {
+      left: 636px;
       top: 112px;
     }
     .save-button-wrapper {
@@ -994,7 +1030,7 @@
       }
       .go-back-screen-button-area {
         top: 94px;
-        left: 325px;
+        left: 342px;
       }      
       .save-button-wrapper {
         text-align: center;
@@ -1161,7 +1197,7 @@
     &.settings {
       .go-back-screen-button-area {
         top: 40px;
-        left: 325px;
+        left: 342px;
       }  
     }    
   }
